+++ services/clients/client.go
@@ -1,0 +1,9 @@
+package clients
+
+import "context"
+
+type Client interface {
+	Start(context.Context) error
+	New() error
+	SetAddress(string, string) error
+}
