+++ services/clients/html/index.html
@@ -4,7 +4,7 @@
     <meta charset="utf-8">
     <title>Timecode Monitor</title>
 
-    <link rel="icon" type="image/png" href="images/favicon.png">
+    <link rel="icon" type="image/png" href="../../../images/favicon.png">
 
     <link rel="preconnect" href="https://fonts.googleapis.com">
     <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
